requirements.txt
fastapi==0.104.1
uvicorn==0.24.0 
sentence-transformers==2.2.2 
tiktoken==0.5.2
numpy==1.26.2 
scipy==1.10.1 
scikit-learn==1.3.0
torchvision 
torchaudio